Tune monospace font size
So Go-Mono looks a little heavy together with the Lora serif font. Tune it down
a notch to make it a little bit more balanced. Unfortunately this makes the
height of inlined monospace text smaller than the main body text, but overall
it's a slight improvement.

Lora + DejaVuSansMono is probably a more balanced look and I may package the
DejaVu font instead in the future.